Set SOURCE_DATE_EPOCH more predictably based on the commit timestamp of the source we're about to build, not any dependencies
If we need the actual "build date" of a given image, that should ultimately come from the build provenance data (in the cases where this will end up getting used).

To put this into context, we want to make sure that if we build an image today, and then rebuild it again (with no source changes) three months from now, possibly even on a newer base image, if it builds reproducibly, we want it to generate the exact same layers (that's the whole point of `SOURCE_DATE_EPOCH`, after all).